Move classNames.js to src/utils subfolder
